minor changes
always return true for deployment validation

add deployment support to ui

update test to remove handling

add integration name to settings, pass settings to kaoto toolbar
